requirements.txt
asgiref==3.8.1
asttokens==2.4.1
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
dj-database-url==2.3.0
Django==5.1.2
executing==2.1.0
gunicorn==23.0.0
h11==0.14.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
wcwidth==0.2.13
whitenoise==6.8.2
django-environ==0.10.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-smtp-ssl==1.0
django-js-asset==2.2.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.2.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.3
django-ckeditor==6.7.0
Pillow==10.0.0